+++ modules/ui-basic/src/ui-basic.module.ts
@@ -1,12 +1,12 @@
 import {NgModule} from "@angular/core";
-import {BrowserModule} from "@angular/platform-browser";
+import {CommonModule} from "@angular/common";
 import {FormsModule, ReactiveFormsModule} from "@angular/forms";
 import {DynamicFormsCoreModule} from "@ng2-dynamic-forms/core";
 import {DynamicFormBasicComponent} from "./dynamic-form-basic.component";
 
 @NgModule({
     imports: [
-        BrowserModule,
+        CommonModule,
         FormsModule,
         ReactiveFormsModule,
         DynamicFormsCoreModule